+++ src/main/java/com/buuz135/industrial/tile/misc/BlackHoleUnitTile.java
@@ -1,10 +1,14 @@
 package com.buuz135.industrial.tile.misc;
 
+import com.buuz135.industrial.proxy.client.ClientProxy;
 import com.buuz135.industrial.proxy.client.infopiece.BlackHoleInfoPiece;
 import com.buuz135.industrial.proxy.client.infopiece.IHasDisplayStack;
 import com.buuz135.industrial.tile.CustomSidedTileEntity;
+import com.buuz135.industrial.utils.ItemStackUtils;
 import com.buuz135.industrial.utils.WorkUtils;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
+import net.minecraft.init.Items;
 import net.minecraft.item.EnumDyeColor;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -14,6 +18,7 @@ import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraft.world.World;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.items.CapabilityItemHandler;
@@ -22,14 +27,18 @@ import net.minecraftforge.items.ItemHandlerHelper;
 import net.ndrei.teslacorelib.gui.BasicTeslaGuiContainer;
 import net.ndrei.teslacorelib.gui.IGuiContainerPiece;
 import net.ndrei.teslacorelib.gui.LockedInventoryTogglePiece;
+import net.ndrei.teslacorelib.gui.ToggleButtonPiece;
 import net.ndrei.teslacorelib.inventory.BoundingRectangle;
 import net.ndrei.teslacorelib.inventory.ColoredItemHandler;
 import net.ndrei.teslacorelib.inventory.LockableItemHandler;
 import net.ndrei.teslacorelib.inventory.SyncProviderLevel;
 import net.ndrei.teslacorelib.items.TeslaWrench;
+import net.ndrei.teslacorelib.netsync.SimpleNBTMessage;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import javax.annotation.Nonnull;
+import java.util.Arrays;
 import java.util.List;
 
 public class BlackHoleUnitTile extends CustomSidedTileEntity implements IHasDisplayStack {
@@ -136,7 +145,35 @@ public class BlackHoleUnitTile extends CustomSidedTileEntity implements IHasDisp
         List<IGuiContainerPiece> list = super.getGuiContainerPieces(container);
         list.add(new LockedInventoryTogglePiece(18 * 2 + 9, 83, this, EnumDyeColor.ORANGE));
         list.add(new BlackHoleInfoPiece(this, 18 * 2 + 8, 25));
+        list.add(new TransferActionButton(136, 84, 13, 13, 0, new ItemStack(Items.BEETROOT_SOUP), "FILL_PLAYER", "fill"));
+        list.add(new TransferActionButton(136 + 18, 84, 13, 13, 0, new ItemStack(Items.BOWL), "EMPTY_PLAYER", "empty"));
         return list;
+    }
+
+
+    @Nullable
+    @Override
+    protected SimpleNBTMessage processClientMessage(@Nullable String messageType, @Nullable EntityPlayerMP player, @NotNull NBTTagCompound compound) {
+        if (player != null && messageType != null) {
+            if (messageType.equalsIgnoreCase("FILL_PLAYER")) {
+                ItemStack stack = this.itemHandler.extractItem(0, 64, true);
+                while (!stack.isEmpty() && player.inventory.addItemStackToInventory(stack)) {
+                    this.itemHandler.extractItem(0, 64, false);
+                    stack = this.itemHandler.extractItem(0, 64, true);
+                }
+                forceSync();
+            }
+            if (messageType.equalsIgnoreCase("EMPTY_PLAYER")) {
+                for (ItemStack itemStack : player.inventory.mainInventory) {
+                    if (!itemStack.isEmpty() && this.itemHandler.insertItem(0, itemStack, true).isEmpty()) {
+                        this.itemHandler.insertItem(0, itemStack.copy(), false);
+                        itemStack.setCount(0);
+                    }
+                }
+                forceSync();
+            }
+        }
+        return super.processClientMessage(messageType, player, compound);
     }
 
     @Override
@@ -283,6 +320,43 @@ public class BlackHoleUnitTile extends CustomSidedTileEntity implements IHasDisp
         @Override
         public int getSlotLimit(int slot) {
             return Integer.MAX_VALUE;
+        }
+    }
+
+    private class TransferActionButton extends ToggleButtonPiece {
+
+        private final ItemStack display;
+        private final String nbtmsg;
+        private final String displayName;
+
+        public TransferActionButton(int left, int top, int width, int height, int hoverOffset, ItemStack display, String nbtmsg, String displayName) {
+            super(left, top, width, height, hoverOffset);
+            this.display = display;
+            this.nbtmsg = nbtmsg;
+            this.displayName = displayName;
+        }
+
+        @Override
+        protected void renderState(BasicTeslaGuiContainer container, int state, BoundingRectangle box) {
+        }
+
+        @Override
+        public void drawBackgroundLayer(BasicTeslaGuiContainer container, int guiX, int guiY, float partialTicks, int mouseX, int mouseY) {
+            super.drawBackgroundLayer(container, guiX, guiY, partialTicks, mouseX, mouseY);
+            container.mc.getTextureManager().bindTexture(ClientProxy.GUI);
+            container.drawTexturedRect(this.getLeft() - 1, this.getTop() - 1, 49, 56, 16, 16);
+            ItemStackUtils.renderItemIntoGUI(display, this.getLeft() + guiX - 2, this.getTop() + guiY - 2, 9);
+        }
+
+        @Override
+        protected void clicked() {
+            BlackHoleUnitTile.this.sendToServer(BlackHoleUnitTile.this.setupSpecialNBTMessage(nbtmsg));
+        }
+
+        @NotNull
+        @Override
+        protected List<String> getStateToolTip(int state) {
+            return Arrays.asList(new TextComponentTranslation("text.industrialforegoing.button.blackhole." + displayName).getFormattedText());
         }
     }
 }